100daysofpython/weather_app/requirements.txt
python-dotenv
geocoder
requests